+++ JointCoopProject/Assets/CYE/Scripts/Items/S/TraitItem.cs
@@ -1,33 +1,33 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+// using System.Collections;
+// using System.Collections.Generic;
+// using UnityEngine;
 
-public class TraitSkill : MonoBehaviour//, IPickable // SkillItem
-{
-    public ItemDataSO _itemData;
+// public class TraitSkill : MonoBehaviour//, IPickable // SkillItem
+// {
+//     public ItemDataSO _itemData;
 
-    #region // Unity Message Function
-    void Awake()
-    {
-        Init();
-    }
-    void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.CompareTag("Player"))
-        {
-            PickedUp();
-        }
-    }
-    #endregion
+//     #region // Unity Message Function
+//     void Awake()
+//     {
+//         Init();
+//     }
+//     void OnTriggerEnter2D(Collider2D collision)
+//     {
+//         if (collision.CompareTag("Player"))
+//         {
+//             PickedUp();
+//         }
+//     }
+//     #endregion
 
-    #region // funciton
-    private void Init()
-    {
+//     #region // funciton
+//     private void Init()
+//     {
 
-    }
-    private void PickedUp()
-    {
+//     }
+//     private void PickedUp()
+//     {
         
-    }
-    #endregion
-}
+//     }
+//     #endregion
+// }
