--- JointCoopProject/Assets/CYE/Scripts/Items/SO/Temp.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ccd76c79b978c4d43a2f0069176e213f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
