--- JointCoopProject/Assets/CYE/Scripts/Items/SO/Temp/New Skill Item SO.asset
@@ -1,24 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5f5df2a0acc0cf949aa035ec2ee4940c, type: 3}
-  m_Name: New Skill Item SO
-  m_EditorClassIdentifier: 
-  _itemID: 1
-  _itemPrefab: {fileID: 2064938242720482303, guid: 901a79b449da06a4cb3bc3fc00a4ffc4, type: 3}
-  _itemName: "\uD14C\uC2A4\uD2B8"
-  _itemPrice: 10
-  _itemIcon: {fileID: 1332096903, guid: 38f3c1bb574f8d0418aad564184fb4b2, type: 3}
-  _itemDesc: "\uD14C\uC2A4\uD2B8\uC6A9 \uC544\uC774\uD15C\uC785\uB2C8\uB2E4."
-  _itemGrade: 0
-  _skillType: 1
-  _itemCooldown: 0
-  _skillDataList: {fileID: 11400000, guid: 270120e841cc04a45a7333b28586b03e, type: 2}
