--- JointCoopProject/Assets/CYE/Scripts/Items/SO/Temp/New Skill Item SO.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 878cffd369ca12c40bbd4de52ac9d4df
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
